// ideqs/src/models/user.tsx
export interface Usuario {
  sucursal: string;
  empresa: string;
  idCliente: number;
  id: number;
  idAws: string;
  user: string;
  activo: number;
  rol: number;
  identificacion: string;
  idSucursal: number;
  correo: string;
  name: string;
  lastname: string;
  pdf: number;
  physical: number;
  mobile: number;
  password?: string;
}